Run checks. (#33)
